feat: make internals public by renaming internal to pkg
- Rename internal/ directory to pkg/ to make Go packages public
- Update all import paths from internal/ to pkg/
- Update CLAUDE.md to reflect new directory structure

This enables other Go programs to import Murmur's packages directly.
