+++ pkg/murmur/providers/awssm/e2e_test.go
@@ -1,21 +1,21 @@
 //go:build e2e
 
-package scwsm_test
+package awssm_test
 
 import (
 	"context"
 	"testing"
 	"time"
 
-	"github.com/busser/murmur/internal/murmur/providers/scwsm"
+	"github.com/busser/murmur/pkg/murmur/providers/awssm"
 )
 
 func TestClient(t *testing.T) {
 
 	// The secrets this test reads were created with Terraform. The code is in
-	// the terraform/layers/scw-secret-manager directory of this repository.
+	// the terraform/layers/aws-secrets-manager directory of this repository.
 
-	client, err := scwsm.New()
+	client, err := awssm.New()
 	if err != nil {
 		t.Fatalf("New() returned an error: %v", err)
 	}
@@ -25,63 +25,34 @@ func TestClient(t *testing.T) {
 		wantVal string
 		wantErr bool
 	}{
+		// References by name.
 		{
 			ref:     "secret-sauce",
 			wantVal: "szechuan",
 			wantErr: false,
 		},
 		{
-			ref:     "secret-sauce#2",
+			ref:     "secret-sauce#AWSCURRENT",
 			wantVal: "szechuan",
 			wantErr: false,
 		},
 		{
-			ref:     "secret-sauce#1",
+			ref:     "secret-sauce#9AF93B18-59D6-4C19-92AC-3F69A115D404",
+			wantVal: "szechuan",
+			wantErr: false,
+		},
+		{
+			ref:     "secret-sauce#v2",
+			wantVal: "szechuan",
+			wantErr: false,
+		},
+		{
+			ref:     "secret-sauce#97DD35A4-DD9B-4E4B-B371-9F2CA4673A41",
 			wantVal: "ketchup",
 			wantErr: false,
 		},
 		{
-			ref:     "fr-par/secret-sauce",
-			wantVal: "szechuan",
-			wantErr: false,
-		},
-		{
-			ref:     "fr-par/secret-sauce#2",
-			wantVal: "szechuan",
-			wantErr: false,
-		},
-		{
-			ref:     "fr-par/secret-sauce#1",
-			wantVal: "ketchup",
-			wantErr: false,
-		},
-		{
-			ref:     "3f34b83f-47a6-4344-bcd4-b63721481cd3",
-			wantVal: "szechuan",
-			wantErr: false,
-		},
-		{
-			ref:     "3f34b83f-47a6-4344-bcd4-b63721481cd3#2",
-			wantVal: "szechuan",
-			wantErr: false,
-		},
-		{
-			ref:     "3f34b83f-47a6-4344-bcd4-b63721481cd3#1",
-			wantVal: "ketchup",
-			wantErr: false,
-		},
-		{
-			ref:     "fr-par/3f34b83f-47a6-4344-bcd4-b63721481cd3",
-			wantVal: "szechuan",
-			wantErr: false,
-		},
-		{
-			ref:     "fr-par/3f34b83f-47a6-4344-bcd4-b63721481cd3#2",
-			wantVal: "szechuan",
-			wantErr: false,
-		},
-		{
-			ref:     "fr-par/3f34b83f-47a6-4344-bcd4-b63721481cd3#1",
+			ref:     "secret-sauce#v1",
 			wantVal: "ketchup",
 			wantErr: false,
 		},
@@ -90,13 +61,40 @@ func TestClient(t *testing.T) {
 			wantVal: "",
 			wantErr: true,
 		},
+
+		// References by ARN.
 		{
-			ref:     "fr-par/does-not-exist",
-			wantVal: "",
-			wantErr: true,
+			ref:     "arn:aws:secretsmanager:eu-west-3:531255069405:secret:secret-sauce-sWcbiZ",
+			wantVal: "szechuan",
+			wantErr: false,
 		},
 		{
-			ref:     "fr-par/does-not-exist#123",
+			ref:     "arn:aws:secretsmanager:eu-west-3:531255069405:secret:secret-sauce-sWcbiZ#AWSCURRENT",
+			wantVal: "szechuan",
+			wantErr: false,
+		},
+		{
+			ref:     "arn:aws:secretsmanager:eu-west-3:531255069405:secret:secret-sauce-sWcbiZ#9AF93B18-59D6-4C19-92AC-3F69A115D404",
+			wantVal: "szechuan",
+			wantErr: false,
+		},
+		{
+			ref:     "arn:aws:secretsmanager:eu-west-3:531255069405:secret:secret-sauce-sWcbiZ#v2",
+			wantVal: "szechuan",
+			wantErr: false,
+		},
+		{
+			ref:     "arn:aws:secretsmanager:eu-west-3:531255069405:secret:secret-sauce-sWcbiZ#97DD35A4-DD9B-4E4B-B371-9F2CA4673A41",
+			wantVal: "ketchup",
+			wantErr: false,
+		},
+		{
+			ref:     "arn:aws:secretsmanager:eu-west-3:531255069405:secret:secret-sauce-sWcbiZ#v1",
+			wantVal: "ketchup",
+			wantErr: false,
+		},
+		{
+			ref:     "arn:aws:secretsmanager:eu-west-3:531255069405:secret:does-not-exist",
 			wantVal: "",
 			wantErr: true,
 		},
