+++ pkg/murmur/providers/jsonmock/client.go
