+++ pkg/murmur/resolve.go
