+++ pkg/murmur/testdata/signal.sh
